Wait for dispatch command to complete if command is quick.
Otherwise, spawn to background.
This hopefully solves the problem where the fifo is blocked when a long
command runs, and it also solves the problem where quality would not
update because we did not wait for dispatch to finish before checking if
quality was updated

Also: Stop die function from erroring.